Move non-template SDF generation methods to .cpp file
